restrukturyzacja projektu
