+++ cpp_bots/cpp_adapter.py
