+++ players/local_player.py
@@ -1,0 +1,9 @@
+from .player import Player
+
+
+class LocalPlayer(Player):
+    def __init__(self, name):
+        self._name = name
+
+    def __str__(self):
+        return self._name
